refactor incident list view
