+++ frontend/src/api/incident.js
@@ -1,26 +1,5 @@
 import moment from "moment";
 import handler from "./apiHandler";
-// import handler from './apiHandler'
-
-// export const createIncident = async (incidentData) => {
-//     return handler.post('/incidents', incidentData);
-// }
-
-// export const getIncident = async (incidentId) => {
-//     return handler.get(`/incidents/${incidentId}`);
-// }
-
-// export const getReporter = async (reporterId) => {
-//     return handler.get(`/reporters/${reporterId}`);
-// }
-
-// export const updateReporter = async (reporterId, reporterData) => {
-//     return handler.put(`/reporters/${reporterId}`, reporterData);
-// }
-
-// export const updateIncident = async (incidentId, incidentData) => {
-//     return handler.put(`/incidents/${incidentId}`, incidentData);
-// }
 
 /*
  * mock API for Incidents
@@ -34,7 +13,6 @@ export const createIncident = async incidentData => {
 };
 
 export const getIncident = async (incidentId) => {
-    // return mockapi.getIncident(incidentId);
     return (await handler.get(`/incidents/${incidentId}`)).data;
 };
 
@@ -83,32 +61,22 @@ export const getIncidents = async (filters, page=1) => {
 };
 
 export const updateIncident = async (incidentId, incidentData) => {
-    // await mockapi.updateIncident(incidentId, incidentData);
-    // return true;
     return (await handler.put(`/incidents/${incidentId}`, incidentData)).data;
 };
 
 export const getReporter = async (reporterId) => {
-    // return mockapi.getReporter(reporterId);
     return (await handler.get(`/reporters/${reporterId}`)).data;
 };
 
 export const updateReporter = async (reporterId, reporterData) => {
-    // return mockapi.updateReporter(reporterId, reporterData);
     return (await handler.put(`/reporters/${reporterId}`, reporterData)).data;
 }
-
 
 export const changeStatus = async (incidentId, status) => {
   return (await handler.get(`/incidents/${incidentId}/status?action=update&type=${status}`)).data;
 };
 
-export const changeSeverity = async (incidentId, severity) => {
-  return mockapi.changeSeverity(incidentId, severity);
-};
-
 export const assignToIncident = async (incidentId, uid) => {
-  // return mockapi.assignToIncident(incidentId, uid);
   return (await handler.get(`/incidents/${incidentId}/assignee?action=change&assignee=${uid}`)).data;
 };
 
@@ -118,7 +86,6 @@ export const removeFromIncident = async (incidentId, uid) => {
 
 export const escallateIncident = async (incidentId) => {
   return (await handler.get(`/incidents/${incidentId}/escalate`)).data;
-  // return mockapi.escallateIncident(incidentId, assigneeId);
 };
 
 export const updateIncidentWorkflow = async (incidentId, workflowType, workflowUpdate) => {
@@ -131,7 +98,6 @@ export const uploadFile = async (incidentId, formData) => {
 }
 
 export const attachMedia = async (incidentId, mediaData) => {
-  // return mockapi.addComment(incidentId, commentData);
   return (await handler.post(`/incidents/${incidentId}/attach_media`, mediaData)).data;
 }
